Denise refactoring, added pixelSource, pixelDma
